backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
alembic

passlib[bcrypt]
python-jose[cryptography]